extracted commands into separate files #59
